+++ modules/global-classes/import-export/import-runner.php
@@ -4,7 +4,7 @@ namespace Elementor\Modules\GlobalClasses\ImportExport;
 
 use Elementor\App\Modules\ImportExport\Runners\Import\Import_Runner_Base;
 use Elementor\App\Modules\ImportExport\Utils as ImportExportUtils;
-use Elementor\Modules\GlobalClasses\Global_Classes_Data;
+use Elementor\Modules\GlobalClasses\Global_Classes_Sanitization;
 use Elementor\Modules\GlobalClasses\Global_Classes_Parser;
 use Elementor\Modules\GlobalClasses\Global_Classes_Repository;
 use Elementor\Plugin;
@@ -38,7 +38,7 @@ class Import_Runner extends Import_Runner_Base {
 			return [];
 		}
 
-		$global_classes = Global_Classes_Data::sanitize_order(
+		$global_classes['order'] = Global_Classes_Sanitization::order(
 			$global_classes['items'] ?? [],
 			$global_classes['order'] ?? []
 		);
